src/bot-bitso-price.py/requirements.txt
python-telegram-bot
bitso-py
